First refactoring commit.
Modified the base code file structure to rely more on header-source architecture for ease of future development
